Add the silver searcher (back) to Ubuntu
There was an old PPA that hasn't been maintained, so I thought that was
not the best idea.
https://launchpad.net/~tomaz-muraus/+archive/the-silver-searcher

There are a few .deb packages made, but they weren't available for
32-bit systems or over SSL.
http://swiftsignal.com/packages/

I couldn't find a good source for an Ubuntu package, so I'm compiling
from source on the master branch of the main github repository.