Remove .zshrc stub creation
Laptop currently generates a .zshrc file if one does not already exist.
However, this file is created with the # DO NOT EDIT BELOW THIS LINE
comment that does not fit with the new thoughtbot/dotfiles convention
of storing changes in .local files instead of above the placeholder
comment. If a user attempts to run laptop and then install dotfiles,
the dotfiles installation will not install .zshrc because of the
existing stub.

This change removes the creation of the .zshrc stub for both mac and
linux scripts.