Install foreman on OS X
This uses installer(1).

On GNU we already install `heroku-toolbelt` using apt, and that package
has a dependency on the `foreman` package.

We install this via installer(1) instead of gem(1) so that it remains
regardless of the development Ruby setup.

We install this via installer(1) instead of from the `Gemfile` so that
it remains regardless of what other developers do on the project (as
described by David Dollar).