Return error code for failure sourcing asdf.sh
source (or `.`) is a built in command and as a result it halts script
execution, and the error code from the command is not returned. This
resulted in the exit code of the script in this case being `0` rather
than the error code of the source command, which is 127 (command not
found).