Switch syntactic to ALE
Notes from Syntastic:


1. Deprecation note

This project is no longer maintained. If you need a syntax checking plugin for Vim you might be interested in Syntastic's spiritual succesor, ALE. Although it shares no code with syntastic and it takes a very different approach to design, ALE can be considered a natural evolution of syntastic in terms of goals and functionality. Check it out, you probably won't be disappointed.